hello_world/requirements.txt
requests
anyio~=3.7.1
argparse~=1.4.0
boto3~=1.33.6
certifi~=2023.11.17
distro~=1.8.0
h11~=0.14.0
httpcore~=1.0.2
httpx~=0.25.2
idna~=3.6
openai~=1.3.7
pydantic~=2.5.2
python-telegram-bot
requests~=2.31.0
sniffio~=1.3.0
urllib3~=1.26.18